Added expected output for all the x86_64 tests.
Each expected output has been checked against the output
of Clang to ensure it looks correct; in some cases we
output excessive allocas (a CodeGen issue that will be
easily fixable) and in other cases we perform direct
loads/stores rather than memcpys (which appears to be
correct).

The main issue remaining is that attributes aren't
currently being added (and signed integers aren't
yet represented in the ABI 'Type' class), so some
tests fail for that reason.