Modified test system line comparison to parse IR to ignore LLVM version-specific instruction differences.
Some versions of LLVM have instructions with different syntax but
the same semantics (at least, from the point of view of ABI
correctness). This change therefore adds some code to parse the IR
and check if the instructions really are different.

This is fairly hacky right now and needs cleanup (and to be made
more comprehensive).